Added a named route with args, allow to go to hub
Added a util class that stores the arguments that have
to be passed. This allows to encapsulate the arguments
passed to the page. To pass arguments, this should always
be done.
Added this named route to main page. This route opens the
specific hub page with it's details. Also changed the lists
in models to dynamic lists since thats what firebase deals in.