Added comment page feed for hubs
This will allow to display comments for hubs and
add new comments. For now this only has buttons
to print comment ids for current hub and add a new
comment. This is for testing. The actual page still
needs to be done. Also did all the backend work
to pass args to this page and function to add comments.
Comment model now also contains user id -> the person
who posted the comment.